+++ NEWS.md
@@ -123,6 +123,7 @@ New library functions
   provide one-argument method that returns a closure.  The old methods of `merge` and
   `merge!` are still available for backward compatibility ([#34296]).
 * The new `isdisjoint` function indicates whether two collections are disjoint ([#34427]).
+* The new `Libc.mkfifo` function wraps `mkfifo` C function on Unix platforms ([#34587]).
 * Add function `ismutable` and deprecate `isimmutable` to check whether something is mutable.([#34652])
 * `include` now accepts an optional `mapexpr` first argument to transform the parsed
   expressions before they are evaluated ([#34595]).